redux refactor
