+++ src/pages/ChangePasswordPage.js
@@ -7,7 +7,7 @@ import * as Yup from 'yup';
 import Container from 'react-bootstrap/Container';
 import { TextInput, Button } from '../components';
 import { showToast } from '../features';
-import { useChangePasswordMutation } from '../services';
+import { useChangePasswordMutation } from '../redux';
 
 export const ChangePasswordPage = () => {
     const navigate = useNavigate();
