+++ src/pages/UserPage.js
@@ -11,7 +11,7 @@ import Alert from 'react-bootstrap/Alert';
 import Card from 'react-bootstrap/Card';
 import Badge from 'react-bootstrap/Badge';
 import { Button, Spinner, Pager } from '../components';
-import { useSearchUsersQuery } from '../services';
+import { useSearchUsersQuery } from '../redux';
 import { ALL_ROLES } from '../utils/auth';
 
 const SORT_OPTIONS = {
