+++ src/redux/index.js
@@ -1,0 +1,2 @@
+export * from './services';
+export * from './store';
