+++ src/redux/services/halcyonApi.js
@@ -1,5 +1,5 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import { config } from '../utils/config';
+import { config } from '../../utils/config';
 
 export const halcyonApi = createApi({
     reducerPath: 'halcyonApi',
